Ignore a NULL directory when trying to clean
output_dir is set to NULL internally when provided explicitly by user to current directory, and in this case it is not empty and should not be cleaned (#857)

+++ R/utils.R
@@ -103,7 +103,7 @@ mark_dirs = function(x) {
 }
 
 clean_empty_dir = function(dir) {
-  if (!dir_exists(dir)) return()
+  if (is.null(dir) || !dir_exists(dir)) return()
   files = list.files(dir, all.files = TRUE, recursive = TRUE)
   if (length(files) == 0) unlink(dir, recursive = TRUE)
 }
